Combine logic into a single file.
